Move remove_auth_from_url and test from Subversion class to pip._internal.utils.misc
